Refactor after merge
